ObjectStore: optionally yield new tree on a fresh tmpfs
This has the benefit of not having a source filesystem in mountinfo
(which will trip up grub2-mkrelpath), but the downside of not being
able to use copy-on-write for our trees.

It may also prove to be faster to run some stages on tmpfs rather
than on a real filesystem, this makes that easier to test.

Signed-off-by: Tom Gundersen <teg@jklm.no>